pagination to TS
