--- src/components/Pagination/index.js
@@ -1,86 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { StyleSheet, TouchableOpacity, View, ViewPropTypes } from 'react-native';
-
-import { colors, vertical, horizontal } from '../../themes';
-
-const styles = StyleSheet.create({
-  container: {
-    position: 'absolute',
-    flexDirection: 'row',
-    marginVertical: vertical.xxSmall,
-    justifyContent: 'center',
-    bottom: 0,
-    width: '100%',
-  },
-  pagination: {
-    width: horizontal.small,
-    height: horizontal.small,
-    borderRadius: 25,
-    marginHorizontal: horizontal.xSmall,
-  },
-});
-
-const Pagination = ({
-  size,
-  paginationIndex,
-  scrollToIndex,
-  paginationDefaultColor,
-  paginationActiveColor,
-  paginationStyle,
-  paginationStyleItem,
-  paginationStyleItemActive,
-  paginationStyleItemInactive,
-  onPaginationSelectedIndex,
-  e2eId,
-}) => {
-  return (
-    <View style={[styles.container, paginationStyle]}>
-      {Array.from({ length: size }).map((_, index) => (
-        <TouchableOpacity
-          testID={`${e2eId}_pagination_${index}`}
-          style={[
-            styles.pagination,
-            paginationStyleItem,
-            paginationIndex === index
-              ? { backgroundColor: paginationActiveColor }
-              : { backgroundColor: paginationDefaultColor },
-            paginationIndex === index ? paginationStyleItemActive : paginationStyleItemInactive,
-          ]}
-          key={index}
-          onPress={() => {
-            scrollToIndex({ index });
-            onPaginationSelectedIndex?.();
-          }}
-        />
-      ))}
-    </View>
-  );
-};
-Pagination.propTypes = {
-  scrollToIndex: PropTypes.func.isRequired,
-  size: PropTypes.number.isRequired,
-  paginationIndex: PropTypes.number,
-  paginationActiveColor: PropTypes.string,
-  paginationDefaultColor: PropTypes.string,
-  paginationStyle: ViewPropTypes.style,
-  paginationStyleItem: ViewPropTypes.style,
-  paginationStyleItemActive: ViewPropTypes.style,
-  paginationStyleItemInactive: ViewPropTypes.style,
-  onPaginationSelectedIndex: PropTypes.func,
-  e2eId: PropTypes.string,
-};
-
-Pagination.defaultProps = {
-  paginationIndex: 0,
-  paginationActiveColor: colors.white,
-  paginationDefaultColor: colors.gray,
-  paginationStyle: {},
-  paginationStyleItem: {},
-  paginationStyleItemActive: {},
-  paginationStyleItemInactive: {},
-  onPaginationSelectedIndex: undefined,
-  e2eId: '',
-};
-
-export default Pagination;
